+++ onboarding.js
@@ -1,0 +1,8 @@
+(function () {
+  "use strict";
+
+  document.getElementById("finish").addEventListener("click", async () => {
+    await window.SeeSettings.setOnboardingState(true);
+    window.close();
+  });
+})();